fixing dataclass and ABC issues.\nmoving to directory for key_pairs\nfinishing EthereumKey
